Introduce project getters module.
