+++ datashuttle/utils/ds_logger.py
@@ -121,9 +121,9 @@ def wrap_variables_for_fancylog(local_vars: dict, cfg: Configs):
     return variables
 
 
-# -------------------------------------------------------------------
+# -----------------------------------------------------------------------------
 # File Snapshots
-# -------------------------------------------------------------------
+# -----------------------------------------------------------------------------
 
 
 def walk_folder(
