+++ datashuttle/utils/ssh.py
@@ -16,9 +16,9 @@ import paramiko
 
 from . import utils
 
-# --------------------------------------------------------------------------------------------------------------------
-# SSH
-# --------------------------------------------------------------------------------------------------------------------
+# -----------------------------------------------------------------------------
+# SSH Setup
+# -----------------------------------------------------------------------------
 
 
 def setup_ssh_key(
@@ -187,6 +187,11 @@ def verify_ssh_central_host(
 def generate_and_write_ssh_key(ssh_key_path: Path) -> None:
     key = paramiko.RSAKey.generate(4096)
     key.write_private_key_file(ssh_key_path.as_posix())
+
+
+# -----------------------------------------------------------------------------
+# Search over SSH
+# -----------------------------------------------------------------------------
 
 
 def search_ssh_central_for_folders(
